Subject [ Feature ] : 加上刪除更新apk邏輯 Message : 進入時，判斷有無更新apk，若有則刪除後重開機

+++ app/src/main/java/com/ufistudio/hotelmediabox/pages/SplashActivity.kt
@@ -3,12 +3,14 @@ package com.ufistudio.hotelmediabox.pages
 import android.content.Intent
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
+import android.util.Log
 import com.ufistudio.hotelmediabox.R
 import com.ufistudio.hotelmediabox.helper.TVController
 import com.ufistudio.hotelmediabox.pages.welcome.WelcomeActivity
 import com.ufistudio.hotelmediabox.repository.data.TVChannel
 import com.ufistudio.hotelmediabox.services.UdpReceiver
 import com.ufistudio.hotelmediabox.utils.FileUtils
+import com.ufistudio.hotelmediabox.utils.MiscUtils
 import com.ufistudio.hotelmediabox.utils.TAG_DEFAULT_APK_NAME
 import java.util.*
 import kotlin.concurrent.schedule
@@ -17,7 +19,7 @@ class SplashActivity : AppCompatActivity() {
 
     private val TAG = SplashActivity::class.java.simpleName
 
-    private var mTVListener:TVController.OnTVListener = object :TVController.OnTVListener{
+    private var mTVListener: TVController.OnTVListener = object : TVController.OnTVListener {
         override fun onChannelChange(tvChannel: TVChannel?) {
         }
 
@@ -37,7 +39,11 @@ class SplashActivity : AppCompatActivity() {
 
     override fun onStart() {
         super.onStart()
-        FileUtils.getFileFromStorage(TAG_DEFAULT_APK_NAME)?.delete()
+        if (FileUtils.getFileFromStorage(TAG_DEFAULT_APK_NAME)?.exists()!!) {
+            FileUtils.getFileFromStorage(TAG_DEFAULT_APK_NAME)?.delete()
+            MiscUtils.reboot(this)
+        }
+
         startService(Intent(this, UdpReceiver::class.java))
     }
 
@@ -53,7 +59,7 @@ class SplashActivity : AppCompatActivity() {
         TVController.releaseListener(mTVListener)
     }
 
-    fun goNextPage(){
+    fun goNextPage() {
         val intent = Intent(this, WelcomeActivity::class.java)
 //        val intent = Intent(this, MainActivity::class.java)
 //        val intent = Intent(this, DVBTestActivity::class.java)
